Fix ferret EMD cost matrix type
On Ubuntu 25.10 GCC 15.2.0 running:

parsecmgmt -a build -p ferret -c gcc-serial

was failing with:

pkgs/apps/ferret/obj/amd64-linux.gcc-serial/src/emd.c:118:12: error: assignment to 'float **' from incompatible pointer type 'double **' [-Wincompatible-pointer-types]

The EMD implementation allocates the cost matrix as double ** and uses double
temporaries throughout. Make emd_state_t::C match that representation instead
of storing it through an incompatible float ** pointer.

+++ pkgs/apps/ferret/src/src/emd.h
@@ -50,7 +50,7 @@ struct emd_node2_t {
 struct emd_state_t {
 	int		n1;			/* SIGNATURES SIZES */
 	int		n2;
-	float		**C;//[MAX_SIG_SIZE1][MAX_SIG_SIZE1]; /* THE COST MATRIX */
+	double		**C;//[MAX_SIG_SIZE1][MAX_SIG_SIZE1]; /* THE COST MATRIX */
 	emd_node2_t	X[MAX_SIG_SIZE1*2];	/* THE BASIC VARIABLES VECTOR */
 
 	/* VARIABLES TO HANDLE _X EFFICIENTLY */
